// src/app/api/generated/model/validacion-de-presentacion.ts
/**
 * AKINE API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ReparoDePresentacion } from './reparo-de-presentacion';


/**
 * Las prestaciones del lote que no se pueden reclamar, con el motivo de cada una. La lista entera y no la primera: hay que poder arreglar todo de una vez.
 */
export interface ValidacionDePresentacion { 
    /**
     * `true` cuando el lote tiene prestaciones y ninguna tiene reparos
     */
    confirmable?: boolean;
    /**
     * Vacio cuando el lote esta listo
     */
    hallazgos?: Array<ReparoDePresentacion>;
    presentacionId?: number;
}